+++ NEWS.md
@@ -1,6 +1,7 @@
 0.37.0 (XXXX-XX-XX)
 * Raised minimum PG version to 11.
-* Add `CDB_RegenerateTable` to regenerate a table.
+* Add `CDB_RegenerateTable` function to regenerate a table.
+* Add `CDB_GetTableQueries` to get the queries of a table (constraints, indices, triggers...).
 * Deprecate creation of new overview tables.
 
 0.36.0 (2020-02-13)